FIX MR: quick fix on merge request gr-d structure * Modified
	- model_initialization.rst: add gr-d structure
	- model_structure.rst: image width to 300
	- _constant.py: black to format
	- forward_db.f90: regenerated
	- md_forward_structure.f90: remove unused variables (l, prd, qd)
	- m_array_creation.f90: fix type conversion int to real
	- test_constant.py: black to format

* Deleted

	- doc.tar.xz
	- doc_old.tar.xz
	- *.kate-swp files

+++ smash/solver/routine/m_array_creation.f90
@@ -26,7 +26,7 @@ module m_array_creation
         
             integer, intent(in) :: stt, stp
             real(sp), intent(in) :: step
-            integer, dimension(ceiling(real(stp - stt, kind=sp) / step)), intent(inout) :: res
+            real(sp), dimension(ceiling(real(stp - stt, kind=sp) / step)), intent(inout) :: res
             
             integer :: i
